pyproject.toml
[tool.poetry]
name = "pythonproject"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.23"}
loguru = "^0.7.2"
alembic = "^1.12.1"
python-dotenv = "^1.0.0"
black = "^23.11.0"
aiosqlite = "^0.19.0"
greenlet = "^3.0.1"
prettytable = "^3.9.0"
pytest-asyncio = "^0.21.1"
respx = "^0.20.2"
tk = "^0.1.0"
customtkinter = "^5.2.1"
httpx = "^0.25.2"
yamldataclassconfig = "^1.5.0"
fastapi = "^0.104.1"
uvicorn = {extras = ["standart"], version = "^0.24.0.post1"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
